+++ tests/option_spec/specs/test_command_specs.py
@@ -55,28 +55,58 @@ describe CommandCase, "array_command_spec":
         command = ["ENV", ["ONE {one}", "TWO {two}"]]
         self.assertDockerLines(command, ["ENV ONE 1", "ENV TWO 2"])
 
-    it "uses complex_ADD_spec if the second value is a dictionary":
+    it "uses complex_ADD_spec if the second value is a dictionary with ADD":
         second_val = {self.unique_val(): self.unique_val()}
         normalised = [Command((self.unique_val(), "one")), Command((self.unique_val(), "two")), Command((self.unique_val(), "three"))]
         normalise = mock.Mock(name="normalise", return_value=normalised)
 
-        command = [self.unique_val(), second_val]
+        command = ["ADD", second_val]
         with mock.patch.object(cs.complex_ADD_spec, "normalise", normalise):
             result = self.spec.normalise(self.meta, command)
         self.assertEqual(result, normalised)
 
+    it "uses complex_ADD_spec if the second value is a dictionary with COPY":
+        second_val = {self.unique_val(): self.unique_val()}
+        normalised = [Command((self.unique_val(), "one")), Command((self.unique_val(), "two")), Command((self.unique_val(), "three"))]
+        normalise = mock.Mock(name="normalise", return_value=normalised)
+
+        command = ["COPY", second_val]
+        with mock.patch.object(cs.complex_COPY_spec, "normalise", normalise):
+            result = self.spec.normalise(self.meta, command)
+        self.assertEqual(result, normalised)
+
 describe CommandCase, "convert_dict_command_spec":
-    it "returns the flattened values from normalising it's spec":
-        val1 = self.unique_val()
-        val2 = self.unique_val()
-        normalised = {self.unique_val(): [val1], self.unique_val(): [val2]}
+    before_each:
+        self.spec = cs.convert_dict_command_spec()
 
-        class spec(object):
-            def normalise(slf, meta, val):
-                return normalised
+    it "uses complex_ADD_spec on the value if the key is ADD":
+        val = {self.unique_val(): self.unique_val()}
+        command = {"ADD": val}
+        normalised = [Command((self.unique_val(), "one")), Command((self.unique_val(), "two")), Command((self.unique_val(), "three"))]
+        normalise = mock.Mock(name="normalise", return_value=normalised)
 
-        result = cs.convert_dict_command_spec(spec()).normalise(self.meta, self.unique_val())
-        self.assertEqual(sorted(result), sorted([val1, val2]))
+        with mock.patch.object(cs.complex_ADD_spec, "normalise", normalise):
+            result = self.spec.normalise(self.meta, command)
+        self.assertEqual(result, normalised)
+
+        normalise.assert_called_once_with(self.meta.at("ADD"), val)
+
+    it "uses complex_COPY_spec on the value if the key is COPY":
+        val = {self.unique_val(): self.unique_val()}
+        command = {"COPY": val}
+        normalised = [Command((self.unique_val(), "one")), Command((self.unique_val(), "two")), Command((self.unique_val(), "three"))]
+        normalise = mock.Mock(name="normalise", return_value=normalised)
+
+        with mock.patch.object(cs.complex_COPY_spec, "normalise", normalise):
+            result = self.spec.normalise(self.meta, command)
+        self.assertEqual(result, normalised)
+
+        normalise.assert_called_once_with(self.meta.at("COPY"), val)
+
+    it "complains if the key isn't ADD or COPY":
+        error = "Commands specified as \[COMMAND, \{options\}\] may only have one option \(either ADD or COPY\)"
+        with self.fuzzyAssertRaisesError(BadSpecValue, error, got="blah", meta=self.meta):
+            self.spec.normalise(self.meta, {"blah": {"content": "blah", "dest": "somewhere"}})
 
 describe CommandCase, "has_a_space validator":
     it "complains if there is no space in the value":
@@ -109,27 +139,6 @@ describe CommandCase, "string_command_spec":
         self.assertEqual(result.command, "{thing}")
         self.assertDockerLines(val, ["FROM {thing}"])
 
-describe CommandCase, "dictionary_command_spec":
-    before_each:
-        self.spec = cs.dictionary_command_spec()
-
-    it "uses complex_ADD_spec on the value":
-        val = self.unique_val()
-        command = {"ADD": val}
-        normalised = [Command((self.unique_val(), "one")), Command((self.unique_val(), "two")), Command((self.unique_val(), "three"))]
-        normalise = mock.Mock(name="normalise", return_value=normalised)
-
-        with mock.patch.object(cs.complex_ADD_spec, "normalise", normalise):
-            result = self.spec.normalise(self.meta, command)
-        self.assertEqual(result, normalised)
-
-        normalise.assert_called_once_with(self.meta.at("ADD"), val)
-
-    it "complains if the key isn't ADD":
-        actual_error = BadSpecValue("Expected the value to be one of the valid choices", choices=("ADD", ), got="blah", meta=self.meta.at("blah"))
-        with self.fuzzyAssertRaisesError(BadSpecValue, meta=self.meta, _errors=[BadSpecValue("Failed to validate", meta=self.meta.at("blah"), _errors=[actual_error])]):
-            self.spec.normalise(self.meta, {"blah": {"content": "blah", "dest": "somewhere"}})
-
 describe CommandCase, "command_spec":
     before_each:
         self.spec = sb.container_spec(Commands, sb.listof(cs.command_spec()))
@@ -140,25 +149,33 @@ describe CommandCase, "command_spec":
         md5 = hashlib.md5(json.dumps({"content": content}).encode('utf-8')).hexdigest()
         md52 = hashlib.md5(json.dumps({"content": {"image": "blah2", "path": "/tmp/stuff"}}, sort_keys=True).encode("utf-8")).hexdigest()
 
+        blah2_image = mock.Mock(name="blah2", image_name="blah2", from_name="somewhere-blah2")
+        blah3_image = mock.Mock(name="blah3", image_name="blah3", from_name="somewhere-blah3")
+
         everything = MergedOptions.using(
               { "mtime": lambda ctxt: 1430660297, "one": 1, "two": 2, "three": 3, "harpoon": self.harpoon, "config_root": "."
               , "images":
-                { "blah": blah_image
-                , "blah2": mock.Mock(name="blah2", image_name="blah2")
+                { "blah2": blah2_image
+                , "blah3": blah3_image
                 }
               }
             )
         meta = Meta(everything, [])
 
         commands = [
-              ["FROM", "{images.blah}"]
+              ["FROM", blah_image]
             , "ADD something /somewhere"
+            , "COPY --from=blah something /somewhere"
             , ["ENV ONE", "{one}"]
             , ["ENV", ["TWO {two}", "THREE {three}"]]
             , ["ADD", {"get": ["blah", "and", "stuff"], "prefix": "/projects"}]
             , {"ADD": {"content": content, "dest": "the_destination"}}
             , {"ADD": {"content": content, "dest": "the_destination2", "mtime": 1530660298}}
             , {"ADD": {"content": {"image": "{images.blah2}", "path": "/tmp/stuff"}, "dest": "the_destination3"}}
+            , {"COPY": {"from": "{images.blah2}", "path": "/tmp/stuff", "to": "copy_destination"}}
+            , {"COPY": {"from": 1, "path": "/tmp/stuff", "to": "copy_destination"}}
+            , ["FROM", "{images.blah3}", "as wat"]
+            , "COPY --from=wat things stuff"
             , "CMD cacafire"
             ]
 
@@ -166,6 +183,7 @@ describe CommandCase, "command_spec":
         self.assertEqual([cmd.as_string for cmd in result.commands]
             , [ "FROM {0}".format(blah_image)
               , "ADD something /somewhere"
+              , "COPY --from=blah something /somewhere"
               , "ENV ONE 1"
               , "ENV TWO 2"
               , "ENV THREE 3"
@@ -175,7 +193,16 @@ describe CommandCase, "command_spec":
               , "ADD {0}-the_destination-mtime(1430660297) the_destination".format(md5)
               , "ADD {0}-the_destination2-mtime(1530660298) the_destination2".format(md5)
               , "ADD {0}-the_destination3-mtime(1430660297) the_destination3".format(md52)
+              , "COPY --from=somewhere-blah2 /tmp/stuff copy_destination"
+              , "COPY --from=1 /tmp/stuff copy_destination"
+              , "FROM somewhere-blah3 as wat"
+              , "COPY --from=wat things stuff"
               , "CMD cacafire"
               ]
             )
 
+        dependents = list(result.dependent_images)
+        self.assertEqual(dependents, [blah_image, blah2_image, blah3_image])
+
+        externals = list(result.external_dependencies)
+        self.assertEqual(externals, [blah_image])